[TACHYON-613] Login Mechanism: Add GetTachyonLoginUser method using LOGIN_MODULEs - refactor structure
