+++ common/src/main/java/tachyon/security/UserInformation.java
@@ -38,110 +38,24 @@ import tachyon.util.PlatformUtils;
 
 /**
  * This class provides methods to determine the login user and connected remote client users.
- * When fetching a login user, it supports Windows, Unix, and Kerberos login modules.
- * When creating a client user, it instantiates a {@link tachyon.security.User} by the user name
- * transmitted by SASL transport from client side.
+ * When fetching a login user of Tachyon service or Tachyon client, it supports Windows, Unix,
+ * and Kerberos login modules.
+ * (TODO) When creating a client user, it instantiates a {@link tachyon.security.User} by the user
+ * name transmitted by SASL transport from client side.
  */
 public class UserInformation {
   private static final Logger LOG = LoggerFactory.getLogger(Constants.LOGGER_TYPE);
 
-  private static final String OS_LOGIN_MODULE_NAME;
-  private static final String OS_PRINCIPAL_CLASS_NAME;
-  private static final boolean WINDOWS = PlatformUtils.OS_NAME.startsWith("Windows");
-  private static final boolean IS_64_BIT = PlatformUtils.PROCESSOR_BIT.contains("64");
-  private static final boolean AIX = PlatformUtils.OS_NAME.equals("AIX");
-
   /** The configuration to use */
   private static final TachyonConf TACHYON_CONF;
-  /** The authentication method to use */
+  /** The authentication method to use. */
   private static AuthenticationFactory.AuthType sAuthType;
-  /** User instance of the login user */
+  /** User instance of the login user in Tachyon process (Tachyon service or Tachyon client) */
   private static User sLoginUser;
 
   static {
-    OS_LOGIN_MODULE_NAME = getOSLoginModuleName();
-    OS_PRINCIPAL_CLASS_NAME = findOsPrincipalClassName();
     TACHYON_CONF = new TachyonConf();
     sAuthType = AuthenticationFactory.getAuthTypeFromConf(TACHYON_CONF);
-  }
-
-  // Return the OS login module class name.
-  private static String getOSLoginModuleName() {
-    if (PlatformUtils.IBM_JAVA) {
-      if (WINDOWS) {
-        return IS_64_BIT ? "com.ibm.security.auth.module.Win64LoginModule"
-            : "com.ibm.security.auth.module.NTLoginModule";
-      } else if (AIX) {
-        return IS_64_BIT ? "com.ibm.security.auth.module.AIX64LoginModule"
-            : "com.ibm.security.auth.module.AIXLoginModule";
-      } else {
-        return "com.ibm.security.auth.module.LinuxLoginModule";
-      }
-    } else {
-      return WINDOWS ? "com.sun.security.auth.module.NTLoginModule"
-          : "com.sun.security.auth.module.UnixLoginModule";
-    }
-  }
-
-  // Return the OS principal class name
-  private static String findOsPrincipalClassName() {
-    String principalClassName = null;
-    if (PlatformUtils.IBM_JAVA) {
-      if (IS_64_BIT) {
-        principalClassName = "com.ibm.security.auth.UsernamePrincipal";
-      } else {
-        if (WINDOWS) {
-          principalClassName = "com.ibm.security.auth.NTUserPrincipal";
-        } else if (AIX) {
-          principalClassName = "com.ibm.security.auth.AIXPrincipal";
-        } else {
-          principalClassName = "com.ibm.security.auth.LinuxPrincipal";
-        }
-      }
-    } else {
-      principalClassName = WINDOWS ? "com.sun.security.auth.NTUserPrincipal"
-          : "com.sun.security.auth.UnixPrincipal";
-    }
-    return principalClassName;
-  }
-
-  static String getOsPrincipalClassName() {
-    return OS_PRINCIPAL_CLASS_NAME;
-  }
-
-  /**
-   * A JAAS configuration that defines the login modules, by which we use to login.
-   */
-  static class TachyonJaasConfiguration extends Configuration {
-    private static final Map<String, String> BASIC_JAAS_OPTIONS = new HashMap<String,String>();
-
-    private static final AppConfigurationEntry OS_SPECIFIC_LOGIN =
-        new AppConfigurationEntry(OS_LOGIN_MODULE_NAME,
-            AppConfigurationEntry.LoginModuleControlFlag.REQUIRED, BASIC_JAAS_OPTIONS);
-
-    private static final AppConfigurationEntry TACHYON_LOGIN =
-        new AppConfigurationEntry(TachyonLoginModule.class.getName(),
-            AppConfigurationEntry.LoginModuleControlFlag.REQUIRED, BASIC_JAAS_OPTIONS);
-
-    // TODO: add Kerberos_LOGIN module
-    // private static final AppConfigurationEntry KERBEROS_LOGIN = ...
-
-    private static final AppConfigurationEntry[] SIMPLE = new
-        AppConfigurationEntry[]{OS_SPECIFIC_LOGIN, TACHYON_LOGIN};
-
-    // TODO: add Kerberos mode
-    // private static final AppConfigurationEntry[] KERBEROS = ...
-
-    @Override
-    public AppConfigurationEntry[] getAppConfigurationEntry(String appName) {
-      if (appName.equalsIgnoreCase(AuthenticationFactory.AuthType.SIMPLE.getAuthName())) {
-        return SIMPLE;
-      } else if (appName.equalsIgnoreCase(AuthenticationFactory.AuthType.KERBEROS.getAuthName())) {
-        // TODO: return KERBEROS;
-        throw new UnsupportedOperationException("Kerberos is not supported currently.");
-      }
-      return null;
-    }
   }
 
   /**
@@ -155,7 +69,10 @@ public class UserInformation {
   }
 
   /**
-   * Get current login user
+   * Get current login user.
+   * This method is called to identify the user who runs Tachyon service or Tachyon client.
+   * When Tachyon client gets a user by this method and connects to Tachyon service,
+   * this user represents the client and is maintained in service.
    * @return the login user
    * @throws IOException if login fails
    */
@@ -170,7 +87,7 @@ public class UserInformation {
    * Login based on the LoginModules
    * @throws IOException if login fails
    */
-  public static void login() throws IOException {
+  private static void login() throws IOException {
     isSecurityEnabled();
     try {
       Subject subject = new Subject();
